requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
beautifulsoup4==4.11.1
certifi==2022.5.18.1
cloudinary==1.29.0
dj-database-url==0.5.0
Django==4.0.5
django-bootstrap4==22.1
django-braces==1.15.0
django-countries==7.3.2
django-crispy-forms==1.14.0
django-heroku==0.3.1
django-model-utils==4.2.0
django-registration-redux==2.10
django-star-ratings==0.9.2
django-taggit==3.0.0
django-tinymce==3.4.0
djangorestframework==3.13.1
gunicorn==20.1.0
Pillow==9.1.1
psycopg2==2.9.3
pytz==2022.1
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
swapper==1.3.0
typing-extensions==4.2.0
urllib3==1.26.9
whitenoise==6.2.0
